# users/users-service.js
const express = require('express');
require('dotenv').config();
const app = express();
const port = 3000;
const swaggerUi = require('swagger-ui-express');
const fs = require('node:fs');
const YAML = require('js-yaml');
const promBundle = require('express-prom-bundle');
const { MongoClient } = require('mongodb');
const crypto = require('node:crypto');

const mongoUri = process.env.DB_URI || 'mongodb://localhost:27017';

function createInMemoryUsersCollection() {
  // Minimal Mongo-like API used by this service (enough for E2E and local dev fallback).
  const users = new Map(); // username -> user doc

  function sanitizeUser(doc) {
    if (!doc) return null;
    // Return a shallow copy to avoid accidental external mutation.
    return { ...doc };
  }

  return {
    async findOne(query) {
      const username = query?.username;
      if (typeof username !== 'string') return null;
      return sanitizeUser(users.get(username) || null);
    },

    async insertOne(doc) {
      if (!doc || typeof doc.username !== 'string') {
        throw new Error('Invalid user document');
      }
      if (users.has(doc.username)) {
        const err = new Error('Duplicate key');
        err.code = 11000;
        throw err;
      }
      users.set(doc.username, sanitizeUser(doc));
      return { acknowledged: true, insertedId: doc.username };
    },

    async updateOne(query, update) {
      const username = query?.username;
      if (typeof username !== 'string') return { matchedCount: 0, modifiedCount: 0 };

      const doc = users.get(username);
      if (!doc) return { matchedCount: 0, modifiedCount: 0 };

      const inc = update?.$inc || {};
      for (const [k, v] of Object.entries(inc)) {
        if (typeof v !== 'number') continue;
        doc[k] = (doc[k] ?? 0) + v;
      }
      users.set(username, sanitizeUser(doc));
      return { matchedCount: 1, modifiedCount: 1 };
    },

    find(_query = {}, options = {}) {
      const projection = options?.projection || null;

      const cursor = {
        _sort: null,

        sort(spec) {
          cursor._sort = spec;
          return cursor;
        },

        async toArray() {
          let arr = Array.from(users.values()).map(sanitizeUser);

          if (cursor._sort && typeof cursor._sort === 'object') {
            const [[field, dir]] = Object.entries(cursor._sort);
            const mul = dir === -1 ? -1 : 1;
            arr.sort((a, b) => {
              const av = a?.[field] ?? 0;
              const bv = b?.[field] ?? 0;
              if (av < bv) return -1 * mul;
              if (av > bv) return 1 * mul;
              return 0;
            });
          }

          if (projection && typeof projection === 'object') {
            arr = arr.map((doc) => {
              const out = {};
              for (const [k, v] of Object.entries(projection)) {
                if (v) out[k] = doc[k];
              }
              return out;
            });
          }

          return arr;
        },
      };

      return cursor;
    },
  };
}

// Default to an in-memory DB so the service is usable even before Mongo connects (or if it never does).
// If Mongo connects successfully, we will replace this with the real collection.
app.locals.usersCollection = createInMemoryUsersCollection();

async function connectToMongo() {
  try {
    const client = new MongoClient(mongoUri);
    await client.connect();
    console.log('Connected to MongoDB');
    const db = client.db('yovi2c_db');
    app.locals.usersCollection = db.collection('users');
  } catch (err) {
    console.error('Failed to connect to MongoDB; using in-memory DB.', err);
  }
}

if (process.env.NODE_ENV !== 'test') {
  connectToMongo();
}

const metricsMiddleware = promBundle({ includeMethod: true });
app.use(metricsMiddleware);

try {
  const swaggerDocument = YAML.load(fs.readFileSync('./openapi.yaml', 'utf8'));
  app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
} catch (e) {
  console.log(e);
}

app.use((req, res, next) => {
  res.setHeader('Access-Control-Allow-Origin', '*');
  res.setHeader('Access-Control-Allow-Methods', 'GET,POST,OPTIONS');
  res.setHeader('Access-Control-Allow-Headers', 'Content-Type');
  if (req.method === 'OPTIONS') return res.sendStatus(204);
  next();
});

app.use(express.json());

const SCRYPT_KEYLEN = 64;

function scryptAsync(password, salt) {
  return new Promise((resolve, reject) => {
    crypto.scrypt(password, salt, SCRYPT_KEYLEN, (err, derivedKey) => {
      if (err) return reject(err);
      resolve(derivedKey);
    });
  });
}

async function hashPassword(password, salt) {
  const derivedKey = await scryptAsync(password, salt);
  return derivedKey.toString('hex');
}

function getValidatedCredentials(req, res) {
  const username = req.body?.username?.trim();
  const password = req.body?.password;

  if (!username) {
    res.status(400).json({ error: 'Username is required' });
    return null;
  }

  if (typeof password !== 'string' || password.length < 6) {
    res.status(400).json({ error: 'Password must be at least 6 characters' });
    return null;
  }

  return { username, password };
}

async function registerHandler(req, res) {
  const credentials = getValidatedCredentials(req, res);
  if (!credentials) return;

  const usersCollection = req.app.locals.usersCollection;
  if (!usersCollection) {
    return res.status(500).json({ error: 'Database not initialized' });
  }

  const { username, password } = credentials;

  try {
    const existingUser = await usersCollection.findOne({ username });
    if (existingUser) {
      return res.status(409).json({ error: 'Username already exists' });
    }

    const passwordSalt = crypto.randomBytes(16).toString('hex');
    const passwordHash = await hashPassword(password, passwordSalt);

    await usersCollection.insertOne({
      username,
      passwordHash,
      passwordSalt,
      createdAt: new Date(),
    });

    return res.status(201).json({ message: 'User registered successfully', username });
  } catch (err) {
    return res.status(500).json({ error: err.message });
  }
}

async function loginHandler(req, res) {
  const credentials = getValidatedCredentials(req, res);
  if (!credentials) return;

  const usersCollection = req.app.locals.usersCollection;
  if (!usersCollection) {
    return res.status(500).json({ error: 'Database not initialized' });
  }

  const { username, password } = credentials;

  try {
    const user = await usersCollection.findOne({ username });
    if (!user || !user.passwordHash || !user.passwordSalt) {
      return res.status(401).json({ error: 'Invalid username or password' });
    }

    const computedHash = await hashPassword(password, user.passwordSalt);
    const validPassword = crypto.timingSafeEqual(
      Buffer.from(user.passwordHash, 'hex'),
      Buffer.from(computedHash, 'hex')
    );

    if (!validPassword) {
      return res.status(401).json({ error: 'Invalid username or password' });
    }

    return res.status(200).json({ message: 'Login successful', username: user.username });
  } catch (err) {
    return res.status(500).json({ error: err.message });
  }
}

//Metodos para el ranking
// ── Registrar resultado de partida ──────────────────────────────────────────

async function gameResultHandler(req, res) {
  // 1. Extraemos y limpiamos los datos
  const rawUsername = req.body?.username?.trim();
  const won = req.body?.won;

  // 2. Validación consistente con el registro
  if (!rawUsername || typeof won !== 'boolean') {
    return res.status(400).json({ error: 'username and won (boolean) are required' });
  }

  // 3. Verificación de la conexión a la base de datos
  const usersCollection = req.app.locals.usersCollection;
  if (!usersCollection) {
    return res.status(500).json({ error: 'Database not initialized' });
  }

  try {
    // 4. Definimos el incremento según el resultado
    const update = won
        ? { $inc: { wins: 1 } }
        : { $inc: { losses: 1 } };

    // 5. CONSTRUCCIÓN SEGURA DE LA QUERY
    // Al crear un objeto nuevo { username: String(rawUsername) },
    // evitamos que SonarQube piense que estamos pasando un objeto malicioso.
    // Usamos String() para asegurar el tipo y que el test lo encuentre por igualdad simple.
    const query = { username: String(rawUsername) };

    const result = await usersCollection.updateOne(query, update);

    // 6. Manejo de resultados
    if (result.matchedCount === 0) {
      return res.status(404).json({ error: 'User not found' });
    }

    return res.status(200).json({ message: 'Result recorded' });
  } catch (err) {
    // Es mejor no devolver el error de sistema directamente por seguridad,
    // pero mantenemos tu estructura original para no romper logs.
    return res.status(500).json({ error: err.message });
  }
}

// ── Obtener ranking ──────────────────────────────────────────────────────────

async function rankingHandler(req, res) {
  const usersCollection = req.app.locals.usersCollection;
  if (!usersCollection) {
    return res.status(500).json({ error: 'Database not initialized' });
  }

  try {
    const players = await usersCollection
        .find(
            {},
            // Solo devolvemos los campos necesarios, nunca el hash ni la sal
            { projection: { _id: 0, username: 1, wins: 1, losses: 1 } }
        )
        .sort({ wins: -1 }) // ordenados de más a menos victorias
        .toArray();

    // Normalizamos: si un jugador nunca ha jugado, wins y losses son 0
    const ranking = players.map((p, index) => ({
      position: index + 1,
      username: p.username,
      wins: p.wins ?? 0,
      losses: p.losses ?? 0,
      // Ratio de victorias: evitamos división por cero
      winRate: (p.wins ?? 0) + (p.losses ?? 0) > 0
          ? Math.round(((p.wins ?? 0) / ((p.wins ?? 0) + (p.losses ?? 0))) * 100)
          : 0,
    }));

    return res.status(200).json({ ranking });
  } catch (err) {
    return res.status(500).json({ error: err.message });
  }
}

// ------------------- Rutas -------------------
app.post('/register', registerHandler);
app.post('/login', loginHandler);

//Nuevas para el ranking
app.post('/game/result', gameResultHandler);  //nuevo
app.get('/ranking', rankingHandler);

// Alias de compatibilidad con versiones previas del frontend.
app.post('/createuser', registerHandler);

if (require.main === module) {
  app.listen(port, () => {
    console.log(`User Service listening at http://localhost:${port}`);
  });
}

module.exports = app;